Extract erase and prepare_database from Observer
